+++ installer/openasar.go
@@ -76,6 +76,9 @@ func (di *DiscordInstall) InstallOpenAsar() error {
 	_ = asarFile.Close()
 
 	if err = os.Rename(asarFile.Name(), path.Join(dir, "app.asar.backup")); err != nil {
+		if errors.Is(err, os.ErrPermission) {
+			return errors.New("The App Management/Full Disk Access permission must be granted to allow VencordInstaller to patch OpenAsar. Make sure Discord isn't running!")
+		}
 		return err
 	}